+++ src/frontend/src/components/molecules/ReplyForm.js
@@ -1,12 +1,9 @@
-import { autoUpdate, flip, offset, shift, useFloating } from '@floating-ui/react';
-import EmojiPicker from 'emoji-picker-react';
 import PropTypes from 'prop-types';
 import React, { useRef, useState } from 'react';
 import AddPhotoAlternateIcon from '@mui/icons-material/AddPhotoAlternate';
-import EmojiEmotionsIcon from '@mui/icons-material/EmojiEmotions';
 import { Box, Button, IconButton, InputAdornment, TextField } from '@mui/material';
-import { useTheme } from '@mui/material/styles';
 import ImagePreview from '../atoms/ImagePreview';
+import FloatingEmojiPicker from 'components/molecules/FloatingEmojiPicker';
 
 const ReplyForm = ({ commentId, onSubmit, onCancel }) => {
   const [text, setText] = useState('');
@@ -14,17 +11,6 @@ const ReplyForm = ({ commentId, onSubmit, onCancel }) => {
   const [imagePreview, setImagePreview] = useState(null);
   const [showEmojiPicker, setShowEmojiPicker] = useState(false);
   const emojiButtonRef = useRef(null);
-
-  const theme = useTheme();
-  const isDarkMode = theme.palette.mode === 'dark';
-
-  const { refs, floatingStyles } = useFloating({
-    open: showEmojiPicker,
-    placement: 'bottom-end',
-    middleware: [offset(4), flip(), shift()],
-    whileElementsMounted: autoUpdate,
-    elements: { reference: emojiButtonRef.current },
-  });
 
   const handleImageChange = (e) => {
     if (e.target.files && e.target.files[0]) {
@@ -49,22 +35,15 @@ const ReplyForm = ({ commentId, onSubmit, onCancel }) => {
 
   const handleKeyDown = (e) => {
     if (e.key === 'Enter' && e.shiftKey) {
-      // Allow Shift+Enter to add a new line
       e.preventDefault();
-      const cursorPosition = e.target.selectionStart; // Get the current cursor position
-      const newText = text.slice(0, cursorPosition) + '\n' + text.slice(cursorPosition); // Insert a new line at the cursor position
+      const cursorPosition = e.target.selectionStart;
+      const newText = text.slice(0, cursorPosition) + '\n' + text.slice(cursorPosition);
       setText(newText);
-
-      // Move the cursor to the correct position after the new line
       setTimeout(() => {
         e.target.setSelectionRange(cursorPosition + 1, cursorPosition + 1);
-
-        // Scroll the TextField to ensure the cursor is visible
-        const inputElement = e.target;
-        inputElement.scrollTop = inputElement.scrollHeight; // Scroll to the bottom of the TextField
+        e.target.scrollTop = e.target.scrollHeight;
       }, 0);
     } else if (e.key === 'Enter') {
-      // Submit the form when Enter is pressed without Shift
       e.preventDefault();
       handleSubmit();
     }
@@ -80,34 +59,21 @@ const ReplyForm = ({ commentId, onSubmit, onCancel }) => {
         onChange={(e) => setText(e.target.value)}
         onKeyDown={handleKeyDown}
         sx={{ mb: 1 }}
-        multiline // Enable multiline input
-        rows={2} // Set initial rows
+        multiline
+        rows={2}
         InputProps={{
           endAdornment: (
             <InputAdornment position="end">
-              <IconButton
-                ref={emojiButtonRef}
-                onClick={() => setShowEmojiPicker((prev) => !prev)}
-                edge="end"
-              >
-                <EmojiEmotionsIcon />
-              </IconButton>
+              <FloatingEmojiPicker
+                onEmojiClick={handleEmojiClick}
+                showEmojiPicker={showEmojiPicker}
+                setShowEmojiPicker={setShowEmojiPicker}
+                emojiButtonRef={emojiButtonRef}
+              />
             </InputAdornment>
           ),
         }}
       />
-      {showEmojiPicker && (
-        <Box
-          ref={refs.setFloating}
-          style={{
-            ...floatingStyles,
-            zIndex: 1000,
-            position: 'absolute',
-          }}
-        >
-          <EmojiPicker onEmojiClick={handleEmojiClick} theme={isDarkMode ? 'dark' : 'light'} />
-        </Box>
-      )}
       {imagePreview && (
         <ImagePreview
           src={imagePreview}
